feat: implement core training infrastructure including custom decoders, trainer logic, and loss modules for SVTRv2 recognition models

+++ openrec/modeling/decoders/smtr_decoder.py
@@ -665,7 +665,7 @@ class SMTRDecoder(nn.Module):
             [targets[2][:, :max_len_curr], targets[5][:, :max_len_curr_pre]],
             1)
         loss1 = F.cross_entropy(answer1_pred.flatten(0, 1),
-                                label.flatten(0, 1),
+                                label.flatten(0, 1).long(),
                                 ignore_index=self.ignore_index,
                                 reduction='mean')
         loss = {'loss': loss1}
